+++ src/backend/access/gist/gistxlog.c
@@ -99,8 +99,7 @@ gistRedoPageUpdateRecord(XLogReaderState *record)
 			itup = (IndexTuple) data;
 			itupsize = IndexTupleSize(itup);
 			if (!PageIndexTupleOverwrite(page, offnum, itup, itupsize))
-				elog(ERROR, "failed to add item to GiST index page, size %d bytes",
-					 (int) itupsize);
+				elog(ERROR, "failed to add item to GiST index page, size %zu bytes", itupsize);
 			data += itupsize;
 			/* should be nothing left after consuming 1 tuple */
 			Assert(data - begin == datalen);
@@ -135,8 +134,7 @@ gistRedoPageUpdateRecord(XLogReaderState *record)
 
 				l = PageAddItem(page, itup, sz, off, false, false);
 				if (l == InvalidOffsetNumber)
-					elog(ERROR, "failed to add item to GiST index page, size %d bytes",
-						 (int) sz);
+					elog(ERROR, "failed to add item to GiST index page, size %zu bytes", sz);
 				off++;
 				ninserted++;
 			}
