+++ src/backend/backup/basebackup_server.c
@@ -176,9 +176,9 @@ bbsink_server_archive_contents(bbsink *sink, size_t len)
 		/* short write: complain appropriately */
 		ereport(ERROR,
 				(errcode(ERRCODE_DISK_FULL),
-				 errmsg("could not write file \"%s\": wrote only %d of %d bytes at offset %u",
+				 errmsg("could not write file \"%s\": wrote only %d of %zu bytes at offset %u",
 						FilePathName(mysink->file),
-						nbytes, (int) len, (unsigned) mysink->filepos),
+						nbytes, len, (unsigned) mysink->filepos),
 				 errhint("Check free disk space.")));
 	}
 
@@ -269,9 +269,9 @@ bbsink_server_manifest_contents(bbsink *sink, size_t len)
 		/* short write: complain appropriately */
 		ereport(ERROR,
 				(errcode(ERRCODE_DISK_FULL),
-				 errmsg("could not write file \"%s\": wrote only %d of %d bytes at offset %u",
+				 errmsg("could not write file \"%s\": wrote only %d of %zu bytes at offset %u",
 						FilePathName(mysink->file),
-						nbytes, (int) len, (unsigned) mysink->filepos),
+						nbytes, len, (unsigned) mysink->filepos),
 				 errhint("Check free disk space.")));
 	}
